requirements.txt
Flask==0.12.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
Jinja2==2.9.6
MarkupSafe==1.0
SQLAlchemy==1.1.13
Tempita==0.5.2
WTForms==2.1
Werkzeug==0.12.2
argparse==1.2.1
blinker==1.4
click==6.7
decorator==4.1.2
itsdangerous==0.24
pbr==3.1.1
python-openid==2.2.5
six==1.10.0
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
wsgiref==0.1.2
